+++ backend/src/modules/user-profile/composition-root.ts
@@ -8,12 +8,24 @@ import type { TypesenseUserProfileServicePort } from "@shared/ports/typesense-us
 import { ProfileRepository } from "./repositories/profile.repository";
 import { PreferenceRepository } from "./repositories/preference.repository";
 import { WorkAreaRepository } from "./repositories/work-area.repository";
+import { EducationRepository } from "./repositories/education.repository";
+import { WorkExperienceRepository } from "./repositories/work-experience.repository";
+import { CertificationRepository } from "./repositories/certification.repository";
+import { SkillRepository } from "./repositories/skill.repository";
 import { ProfileService } from "./services/profile.service";
 import { PreferenceService } from "./services/preference.service";
 import { WorkAreaService } from "./services/work-area.service";
+import { EducationService } from "./services/education.service";
+import { WorkExperienceService } from "./services/work-experience.service";
+import { CertificationService } from "./services/certification.service";
+import { SkillService } from "./services/skill.service";
 import { ProfileController } from "./controllers/profile.controller";
 import { PreferenceController } from "./controllers/preference.controller";
 import { WorkAreaController } from "./controllers/work-area.controller";
+import { EducationController } from "./controllers/education.controller";
+import { WorkExperienceController } from "./controllers/work-experience.controller";
+import { CertificationController } from "./controllers/certification.controller";
+import { SkillController } from "./controllers/skill.controller";
 import { createProfileGuards } from "./guards/profile.guards";
 import { createTypesenseUserProfileIndexerWorker } from "./workers/typesense-user-profile-indexer.worker";
 
@@ -35,6 +47,10 @@ export function createUserProfileModule(deps: UserProfileModuleDeps) {
   const repository = deps.profileRepository;
   const preferenceRepository = new PreferenceRepository();
   const workAreaRepository = new WorkAreaRepository();
+  const educationRepository = new EducationRepository();
+  const workExperienceRepository = new WorkExperienceRepository();
+  const certificationRepository = new CertificationRepository();
+  const skillRepository = new SkillRepository();
 
   const service = new ProfileService(
     repository,
@@ -51,10 +67,31 @@ export function createUserProfileModule(deps: UserProfileModuleDeps) {
     preferenceRepository,
     repository,
   );
+  const educationService = new EducationService(
+    educationRepository,
+    repository,
+  );
+  const workExperienceService = new WorkExperienceService(
+    workExperienceRepository,
+    repository,
+  );
+  const certificationService = new CertificationService(
+    certificationRepository,
+    repository,
+  );
+  const skillService = new SkillService(skillRepository, repository);
 
   const controller = new ProfileController(service, deps.userOrgsQuery);
   const preferenceController = new PreferenceController(preferenceService);
   const workAreaController = new WorkAreaController(workAreaService);
+  const educationController = new EducationController(educationService);
+  const workExperienceController = new WorkExperienceController(
+    workExperienceService,
+  );
+  const certificationController = new CertificationController(
+    certificationService,
+  );
+  const skillController = new SkillController(skillService);
 
   const guards = createProfileGuards({ profileRepository: repository });
 
@@ -66,6 +103,10 @@ export function createUserProfileModule(deps: UserProfileModuleDeps) {
     controller,
     preferenceController,
     workAreaController,
+    educationController,
+    workExperienceController,
+    certificationController,
+    skillController,
     guards,
     repository,
     service,
