+++ backend/src/modules/user-profile/ports/certification-repository.port.ts
@@ -1,0 +1,18 @@
+import type {
+  Certification,
+  NewCertification,
+} from "@/validations/certifications.validation";
+
+export interface CertificationRepositoryPort {
+  linkCertification(
+    userProfileId: number,
+    certificationData: NewCertification,
+  ): Promise<Certification>;
+
+  unlinkCertification(
+    userProfileId: number,
+    certificationId: number,
+  ): Promise<boolean>;
+
+  searchCertifications(query: string): Promise<Certification[]>;
+}
