+++ backend/src/modules/user-profile/ports/education-repository.port.ts
@@ -1,0 +1,18 @@
+import type {
+  Education,
+  InsertEducation,
+} from "@/validations/educations.validation";
+
+export interface EducationRepositoryPort {
+  batchAddEducations(
+    userProfileId: number,
+    data: Omit<InsertEducation, "userProfileId">[],
+  ): Promise<Education[]>;
+
+  updateEducation(
+    educationId: number,
+    data: Partial<Omit<InsertEducation, "userProfileId">>,
+  ): Promise<boolean>;
+
+  deleteEducation(educationId: number): Promise<boolean>;
+}
