+++ backend/src/modules/user-profile/ports/profile-service.port.ts
@@ -7,21 +7,7 @@ import type {
   UserProfile,
   UserWithProfile,
 } from "@/validations/userProfile.validation";
-import type {
-  Education,
-  InsertEducation,
-} from "@/validations/educations.validation";
-import type {
-  InsertWorkExperience,
-  WorkExperience,
-} from "@/validations/workExperiences.validation";
-import type {
-  Certification,
-  NewCertification,
-} from "@/validations/certifications.validation";
-import type { NewSkill, Skill } from "@/validations/skills.validation";
 import type { PaginationMeta } from "@shared/types";
-import { SavedJobs } from "@/validations/user.validation";
 import {
   ProfilePictureFile,
   ResumeFile,
@@ -87,99 +73,17 @@ export interface ProfileServicePort {
     intent: "seeker" | "employer",
   ): Promise<Result<void, AppError>>;
 
-  getSavedJobsForUser(
-    userId: number,
-    page?: number,
-    limit?: number,
-  ): Promise<
-    Result<{ items: SavedJobs[]; pagination: PaginationMeta }, AppError>
-  >;
-
-  saveJobForCurrentUser(
-    userId: number,
-    jobId: number,
-  ): Promise<Result<{ success: boolean }, AppError>>;
-
-  isJobSavedByUser(
-    userId: number,
-    jobId: number,
-  ): Promise<Result<boolean, AppError>>;
-
-  unsaveJobForCurrentUser(
-    userId: number,
-    jobId: number,
-  ): Promise<Result<{ success: boolean }, AppError>>;
-
-  // Education CRUD
-  batchAddEducations(
-    userId: number,
-    data: Omit<InsertEducation, "userProfileId">[],
-  ): Promise<Result<Education[], AppError>>;
-
-  updateEducation(
-    educationId: number,
-    data: Partial<Omit<InsertEducation, "userProfileId">>,
-  ): Promise<Result<boolean, AppError>>;
-
-  deleteEducation(educationId: number): Promise<Result<boolean, AppError>>;
-
-  // Work Experience CRUD
-  batchAddWorkExperiences(
-    userId: number,
-    data: Omit<InsertWorkExperience, "userProfileId">[],
-  ): Promise<Result<WorkExperience[], AppError>>;
-
-  updateWorkExperience(
-    workExperienceId: number,
-    data: Partial<Omit<InsertWorkExperience, "userProfileId">>,
-  ): Promise<Result<boolean, AppError>>;
-
-  deleteWorkExperience(
-    workExperienceId: number,
-  ): Promise<Result<boolean, AppError>>;
-
-  // Certification link/unlink/search
-  linkCertification(
-    userId: number,
-    certificationData: NewCertification,
-  ): Promise<Result<Certification, AppError>>;
-
-  unlinkCertification(
-    userId: number,
-    certificationId: number,
-  ): Promise<Result<boolean, AppError>>;
-
-  searchCertifications(
-    query: string,
-  ): Promise<Result<Certification[], AppError>>;
-
-  // Skill link/unlink/search
-  linkSkill(
-    userId: number,
-    skillData: NewSkill,
-  ): Promise<Result<Skill, AppError>>;
-
-  unlinkSkill(
-    userId: number,
-    skillId: number,
-  ): Promise<Result<boolean, AppError>>;
-
-  searchSkills(query: string): Promise<Result<Skill[], AppError>>;
-
-  // Profile picture upload
   uploadProfilePicture(
     userId: number,
     file: ProfilePictureFile | undefined,
     correlationId: string,
   ): Promise<Result<{ message: string }, AppError>>;
 
-  // Resume upload
   uploadResume(
     userId: number,
     file: ResumeFile | undefined,
     correlationId: string,
   ): Promise<Result<{ message: string }, AppError>>;
 
-  // Resume delete
   deleteResume(userId: number): Promise<Result<{ message: string }, AppError>>;
 }
