+++ backend/src/modules/user-profile/ports/work-experience-repository.port.ts
@@ -1,0 +1,18 @@
+import type {
+  InsertWorkExperience,
+  WorkExperience,
+} from "@/validations/workExperiences.validation";
+
+export interface WorkExperienceRepositoryPort {
+  batchAddWorkExperiences(
+    userProfileId: number,
+    data: Omit<InsertWorkExperience, "userProfileId">[],
+  ): Promise<WorkExperience[]>;
+
+  updateWorkExperience(
+    workExperienceId: number,
+    data: Partial<Omit<InsertWorkExperience, "userProfileId">>,
+  ): Promise<boolean>;
+
+  deleteWorkExperience(workExperienceId: number): Promise<boolean>;
+}
