+++ backend/src/modules/user-profile/ports/work-experience-service.port.ts
@@ -1,0 +1,22 @@
+import type { Result } from "@shared/result";
+import type { AppError } from "@shared/errors";
+import type {
+  InsertWorkExperience,
+  WorkExperience,
+} from "@/validations/workExperiences.validation";
+
+export interface WorkExperienceServicePort {
+  batchAddWorkExperiences(
+    userId: number,
+    data: Omit<InsertWorkExperience, "userProfileId">[],
+  ): Promise<Result<WorkExperience[], AppError>>;
+
+  updateWorkExperience(
+    workExperienceId: number,
+    data: Partial<Omit<InsertWorkExperience, "userProfileId">>,
+  ): Promise<Result<boolean, AppError>>;
+
+  deleteWorkExperience(
+    workExperienceId: number,
+  ): Promise<Result<boolean, AppError>>;
+}
