+++ backend/tests/unit/modules/user-profile/repositories/profile.repository.test.ts
@@ -1,30 +1,47 @@
-import { Mock, vi } from "vitest";
+import { describe, it, expect, vi } from "vitest";
 
-vi.mock("@shared/db/connection", () => ({
-  db: {
-    transaction: vi.fn(),
-  },
-}));
+import { SavedJobService } from "@/modules/applications/services/saved-job.service";
+import type { SavedJobRepositoryPort } from "@/modules/applications/ports/saved-job-repository.port";
 
-import { db } from "@shared/db/connection";
-import { ProfileRepository } from "@/modules/user-profile/repositories/profile.repository";
+function createMockRepository(
+  overrides: Partial<SavedJobRepositoryPort> = {},
+): SavedJobRepositoryPort {
+  return {
+    getSavedJobsForUser: vi.fn(),
+    saveJobForUser: vi.fn().mockResolvedValue({ success: true }),
+    isJobSavedByUser: vi.fn(),
+    countSavedJobs: vi.fn().mockResolvedValue(0),
+    getSavedJobIdsForJobs: vi.fn(),
+    unsaveJobForUser: vi.fn(),
+    ...overrides,
+  };
+}
 
-describe("ProfileRepository.saveJobForUser - saved jobs limit", () => {
-  it("throws DatabaseError when saved jobs >= 50", async () => {
-    const txMock = {
-      $count: vi.fn().mockResolvedValue(50),
-    };
-
-    (db.transaction as Mock).mockImplementation(async (cb: any) => {
-      return cb(txMock);
+describe("SavedJobService.saveJobForCurrentUser - saved jobs limit", () => {
+  it("returns failure when saved jobs >= 50", async () => {
+    const repo = createMockRepository({
+      countSavedJobs: vi.fn().mockResolvedValue(50),
     });
+    const service = new SavedJobService(repo);
 
-    const repo = new ProfileRepository();
+    const result = await service.saveJobForCurrentUser(1, 123);
 
-    await expect(repo.saveJobForUser(1, 123)).rejects.toThrowError(
+    expect(result.isFailure).toBe(true);
+    expect(result.error?.message).toMatch(
       /Saved jobs limit reached. You can save up to 50 jobs./,
     );
+    expect(repo.saveJobForUser).not.toHaveBeenCalled();
+  });
 
-    expect(txMock.$count).toHaveBeenCalled();
+  it("saves successfully when under the limit", async () => {
+    const repo = createMockRepository({
+      countSavedJobs: vi.fn().mockResolvedValue(49),
+    });
+    const service = new SavedJobService(repo);
+
+    const result = await service.saveJobForCurrentUser(1, 123);
+
+    expect(result.isSuccess).toBe(true);
+    expect(repo.saveJobForUser).toHaveBeenCalledWith(1, 123);
   });
 });
